Demonstration of prototype
The application gets the users location via the HTML5 Geolocation API, then uses the HERE API to work out which mesh block (a polygon representing a geographic area in the police data) the user is in.  Once the mesh block ID is determined, a polygon is overlaid onto the user map - this is stored in a MySQL database and retrieved by mesh block ID.  The parameters of this polygon are changed depending on the police crimes data, at the moment stored locally in a MySQL database.